pyproject.toml
[tool.poetry]
name = "linkedin-summarizer"
version = "0.1.0"
description = ""
authors = ["SayaleeDamle <sayalee.damle@onepointltd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.325"
openai = "^0.28.1"
pydantic = "^1.9"
python-dotenv = "^1.0.0"
chainlit = "^0.7.301"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
black = "^23.10.1"
farm-haystack = "^1.21.2"
pyautogen = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
